fix syntax error

+++ samgraph/common/cpu/cpu_loops.cc
@@ -58,7 +58,7 @@ void DoCPUSample(TaskPtr task) {
 
   for (int i = last_layer_idx; i >= 0; i--) {
     Timer t0;
-    const int fanout = fanouts[i];
+    const size_t fanout = fanouts[i];
     const IdType *input = static_cast<const IdType *>(cur_input->Data());
     const size_t num_input = cur_input->Shape()[0];
     LOG(DEBUG) << "CPUSample: begin sample layer " << i;
